feat: move classes into their separate files.
